+++ src/main/java/io/zipcoder/casino/utilities/Input.java
@@ -1,39 +1,62 @@
 package io.zipcoder.casino.utilities;
 
+
 import java.io.InputStream;
 import java.io.PrintStream;
 import java.util.Scanner;
 
-public class Input {
-    public static Scanner scan = new Scanner(System.in);
+/**
+ * You are advised against modifying this class.
+ */
+public final class Input {
+    private final Scanner input;
+    private final PrintStream output;
 
-    public static String getStringInput(String prompt) {
-        String userInput;
-        System.out.print(prompt);
-        if (scan.hasNextLine()){
-           userInput = scan.next();
-        } else {
-            userInput = getStringInput("Incorrect input, try again:  ");
-        }
-        return userInput;
+    public Input(InputStream in, PrintStream out) {
+        this.input = new Scanner(in);
+        this.output = out;
     }
 
-    public static Integer getIntInput(String prompt) {
 
-        System.out.print(prompt);
-        Integer userInput = scan.nextInt();
-        return userInput;
+    public void print(String val, Object... args) {
+        output.format(val, args);
     }
 
-    public static Double getDoubleInput(String prompt) {
-        Double userInput;
-        System.out.print(prompt);
-        if(scan.hasNextDouble()){
-            userInput = scan.nextDouble();
+    public void println(String val, Object... vals) {
+        print(val + "\n", vals);
+    }
+
+    public String getStringInput(String prompt, Object... args) {
+        print(prompt, args);
+        return input.nextLine();
+    }
+
+    public Double getDoubleInput(String prompt, Object... args) {
+        String stringInput = getStringInput(prompt, args);
+        try {
+            Double doubleInput = Double.parseDouble(stringInput);
+            return doubleInput;
+        } catch (NumberFormatException nfe) { // TODO - Eliminate recursive nature
+            println("[ %s ] is an invalid user input!", stringInput);
+            println("Try inputting a numeric value!");
+            return getDoubleInput(prompt, args);
         }
-        else{
-            userInput = getDoubleInput("Incorrect input, try again:  ");
+    }
+
+    public Long getLongInput(String prompt, Object... args) {
+        String stringInput = getStringInput(prompt, args);
+        try {
+            Long longInput = Long.parseLong(stringInput);
+            return longInput;
+        } catch (NumberFormatException nfe) { // TODO - Eliminate recursive nature
+            println("[ %s ] is an invalid user input!", stringInput);
+            println("Try inputting an integer value!");
+            return getLongInput(prompt, args);
         }
-        return userInput;
+    }
+
+    public Integer getIntegerInput(String prompt, Object... args) {
+        return getLongInput(prompt, args).intValue();
     }
 }
+
