+++ src/main/java/io/zipcoder/casino/utilities/Output.java
@@ -1,12 +1,10 @@
 package io.zipcoder.casino.utilities;
 
-import javax.xml.bind.annotation.adapters.CollapsedStringAdapter;
-
 public class Output {
     Integer selection;
     Double walletAmount;
     String name;
-
+    private static Input input = new Input(System.in,System.out);
     /**
      * Welcome
      * Get user name
@@ -32,33 +30,34 @@ public class Output {
     }
 
     public String getPlayerName(){
-        name = Input.getStringInput("Enter your name:  \n");
+        name = input.getStringInput("Enter your name:  \n");
         return name;
     }
 
     public Boolean isGambler(){
+
         String line1 = "~~~~~~~~~~~ Press 1 to gamble today! ~~~~~~~~~~~~\n";
         String line2 = "~~~~ Press any other number to play it safe! ~~~~\n";
-        selection = Input.getIntInput(line1 + line2);
-        return selection == 1 ? true:false;
+        selection = input.getIntegerInput(line1 + line2);
+        return selection == 1.0 ? true:false;
     }
 
     public Double getWalletAmount(){
-         walletAmount = Input.getDoubleInput("Deposit money into you wallet to gamble\nEnter amount: ");
+         walletAmount = input.getDoubleInput("Deposit money into you wallet to gamble\nEnter amount: ");
          return walletAmount;
     }
 
     public Integer allGameOptions(){
         String line1 = "Which game would you like to play!?\n";
         String line2 ="1:BlackJack  2:GoFish  3:Cho-Han  4:Craps  5:Exit Casino\n";
-        selection = Input.getIntInput(line1 + line2);
+        selection = input.getIntegerInput(line1 + line2);
         return selection;
     }
 
     public Integer nonGamblingOptions(){
         String line1 = "Which game would you like to play!?\n";
         String line2 ="1:GoFish  2:Cho-Han  3:Become a gambler  4:Exit Casino\n";
-        selection = Input.getIntInput(line1 + line2);
+        selection = input.getIntegerInput(line1 + line2);
         return selection;
     }
