+++ src/main/java/io/zipcoder/casino/utilities/Switchboard.java
@@ -6,7 +6,10 @@ import io.zipcoder.casino.Games.DiceGames.ChoHan;
 import io.zipcoder.casino.Games.DiceGames.Craps;
 import io.zipcoder.casino.Players.*;
 
+import java.util.Collection;
+
 public class Switchboard {
+
     Output output = new Output();
     String name;
     Boolean isGambler;
@@ -24,8 +27,6 @@ public class Switchboard {
             }
         output.goodBye();
         }
-
-
 
 
     public Player setPlayer(){
